requirements.txt
openai==1.30.1
streamlit==1.34.0
streamlit-feedback==0.1.3
vanna==0.5.5
pandas==2.2.2
pyodbc==5.1.0
plotly==5.22.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
Unidecode==1.3.8
numpy==1.26.4
requests==2.32.2
regex==2024.5.15
chroma-hnswlib==0.7.3
chromadb==0.5.0
Flask==3.0.3
flask-sock==0.7.0
Flask-Cors==4.0.1
# pysqlite3-binary==0.5.2.post3
gunicorn==22.0.0
scikit-learn
statsmodels==0.14.2
python-dotenv